added AreaCalculator; replaced double with BigDecimal;

+++ src/swed/it/academy/project/Square.java
@@ -1,8 +1,10 @@
 package swed.it.academy.project;
 
+import java.math.BigDecimal;
+
 public class Square extends Shape
 {
-    public Square(double edgeLength)
+    public Square(BigDecimal edgeLength)
     {
         super(edgeLength, edgeLength);
     }
